member3_broadcast/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "member3-realtime-websocket"
version = "0.1.0"
description = "Member 3 real-time WebSocket broadcast service for campus intelligence"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "python-socketio[asgi]>=5.13",
  "httpx>=0.27",
  "PyJWT>=2.9",
  "pydantic>=2.7",
  "pydantic-settings>=2.4",
]

[project.optional-dependencies]
test = [
  "pytest>=8.3",
  "pytest-asyncio>=0.23",
]

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-q"
